# docs/submission/openapi_service_control.yaml
openapi: 3.0.3
info:
  title: AI Service Control API
  version: "0.5.0"
  description: >
    Go Echo contract for the KHU AI application automation Agent. The core flow
    joins ApplicationProfile and ResourceRecommendation messages, authorizes the
    built-in AIApplicationAutomationAgent through Agent Registry capability and
    bounded Action policy, decides DEPLOY, REJECT, or RETRY, applies deterministic
    Go Guard checks, and emits a platform-neutral Desired Deployment Spec inside
    deployment.create.request. Deployment execution and platform-specific Manifest
    conversion remain external responsibilities. Legacy ControlRun, runtime Agent
    execution, and post-deployment controls are retained as optional PoC tools.
servers:
  - url: http://127.0.0.1:18080
paths:
  /api/v1/agent-control/external-flows:
    post:
      summary: Validate and join an external input pair atomically
      description: Matching correlation_id, trace_id and profile_id are required. Identical replays return the existing Flow; conflicting reuse is rejected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [application_context, resource_recommendation]
              properties:
                application_context: {$ref: '#/components/schemas/ApplicationContextEnvelope'}
                resource_recommendation: {$ref: '#/components/schemas/ResourceRecommendationEnvelope'}
                decision_agent: {type: string}
      responses:
        '202':
          description: Joined Flow and guarded Agent decision
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AgentControlFlow'}
        '400': {description: Invalid or conflicting external pair}
...
  /api/v1/agent-control/flows/{correlation_id}/appdeploy/{action}:
    post:
      summary: Prepare or explicitly submit approved initial Revision 1
      description: Requires an App Registry identity match. Only one node and one replica are supported. Registered AppSpec settings and AppDeployer placement are used; inference settings and GPU VRAM minima are not enforced. Submission attempts are journaled and uncertain results are not automatically retried.
      parameters:
        - {name: correlation_id, in: path, required: true, schema: {type: string}}
        - {name: action, in: path, required: true, schema: {type: string, enum: [prepare, submit]}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [app_version_id]
              properties:
                app_version_id: {type: string}
                accept_projection_limits: {type: boolean, description: Must be true for submission}
      responses:
        '200':
          description: Prepared request or durable delivery receipt
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FlowDelivery'}
        '400': {description: Invalid request}
        '409': {description: Submission disabled, invalid binding, unapproved Flow or unconfirmed upstream response}
  /api/v1/agent-control/flows/{correlation_id}/appdeploy:
    get:
      summary: Query real deployment state and raw metrics for a recorded delivery
      parameters:
        - {name: correlation_id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: Receipt with status and available metrics; no synthetic optimization feedback
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FlowDelivery'}
        '409': {description: Missing receipt, changed destination or upstream query failure}
...
components:
  securitySchemes:
    AutonomyAdminToken:
      type: http
      scheme: bearer
      bearerFormat: opaque admin token
...
  schemas:
...
    CommonMessageEndpoint:
      type: object
      required: [system, component]
      properties:
        system: {type: string}
        component: {type: string}
    CommonMessageEnvelope:
      type: object
      required:
        - contract_version
        - message_id
        - message_type
        - occurred_at
        - correlation_id
        - trace_id
        - source
        - target
      properties:
        contract_version: {type: string, enum: ["1.0"]}
        message_id: {type: string}
        message_type: {type: string}
        occurred_at: {type: string, format: date-time}
        correlation_id: {type: string}
        trace_id: {type: string}
        causation_id: {type: string}
        source:
          $ref: "#/components/schemas/CommonMessageEndpoint"
        target:
          $ref: "#/components/schemas/CommonMessageEndpoint"
...
    ApplicationContextEnvelope:
      allOf:
        - $ref: "#/components/schemas/CommonMessageEnvelope"
        - type: object
          required: [data]
          properties:
            message_type:
              type: string
              enum: [application.context.created]
            data:
              type: object
              required: [application_profile, model_recommendation]
              properties:
                application_profile:
                  $ref: "#/components/schemas/ApplicationProfile"
                model_recommendation:
                  $ref: "#/components/schemas/ApplicationModelRecommendation"
    ApplicationProfile:
      type: object
      required: [profile_id, app_id, app_version, requirements]
      properties:
        profile_id: {type: string}
        app_id: {type: string}
        app_version: {type: string}
        artifact:
          type: object
          properties:
            type: {type: string}
            uri: {type: string}
            entrypoint:
              type: array
              items: {type: string}
        workload:
          type: object
          properties:
            task_type: {type: string}
            request_pattern: {type: string}
            expected_rps: {type: number, minimum: 0}
        requirements:
          $ref: "#/components/schemas/ApplicationRequirements"
        analysis:
          type: object
          properties:
            confidence: {type: number, minimum: 0, maximum: 1}
            assumptions: {type: array, items: {type: string}}
            missing_fields: {type: array, items: {type: string}}
            warnings: {type: array, items: {type: string}}
    ApplicationRequirements:
      type: object
      required: [compute, accelerator, deployment]
      properties:
        compute:
          type: object
          required: [cpu_cores_min, memory_mib_min, storage_gib_min]
          properties:
            cpu_cores_min: {type: integer, minimum: 1}
            memory_mib_min: {type: integer, minimum: 1}
            storage_gib_min: {type: integer, minimum: 0}
        accelerator:
          type: object
          required: [required]
          properties:
            required: {type: boolean}
            type: {type: string, example: GPU}
            count_min: {type: integer, minimum: 0}
            memory_mib_min_per_device: {type: integer, minimum: 0}
        deployment:
          type: object
          required: [replicas_min, replicas_max, isolation]
          properties:
            replicas_min: {type: integer, minimum: 1}
            replicas_max: {type: integer, minimum: 1}
            isolation: {type: string}
        slo:
          type: object
          properties:
            latency_p95_ms_max: {type: number, minimum: 0}
            throughput_rps_min: {type: number, minimum: 0}
        cost:
          type: object
          properties:
            currency: {type: string}
            cost_per_hour_max: {type: number, minimum: 0}
    ApplicationModelRecommendation:
      type: object
      required: [selected_model, inference_configuration]
      properties:
        recommendation_id: {type: string}
        selected_model:
          type: object
          required: [model_id, model_version, source]
          properties:
            model_id: {type: string}
            model_version: {type: string}
            source: {type: string}
        inference_configuration:
          type: object
          required:
            - runtime_engine
            - precision
            - max_batch_size
            - max_concurrency
            - tensor_parallel_size
            - replicas
          properties:
            runtime_engine: {type: string}
            precision: {type: string}
            max_batch_size: {type: integer, minimum: 1}
            max_concurrency: {type: integer, minimum: 1}
            tensor_parallel_size: {type: integer, minimum: 1}
            replicas: {type: integer, minimum: 1}
    ResourceRecommendationEnvelope:
      allOf:
        - $ref: "#/components/schemas/CommonMessageEnvelope"
        - type: object
          required: [data]
          properties:
            message_type:
              type: string
              enum: [resource.recommendation.created]
            data:
              type: object
              required: [resource_recommendation]
              properties:
                resource_recommendation:
                  $ref: "#/components/schemas/ResourceRecommendation"
    ResourceRecommendation:
      type: object
      required:
        - recommendation_id
        - profile_id
        - snapshot_id
        - status
        - selected_candidate_id
        - candidates
      properties:
        recommendation_id: {type: string}
        profile_id: {type: string}
        snapshot_id: {type: string}
        status: {type: string}
        selected_candidate_id: {type: string}
        candidates:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ResourceCandidate"
    DeploymentStatusEnvelope:
      allOf:
        - $ref: "#/components/schemas/CommonMessageEnvelope"
        - type: object
          required: [data]
          properties:
            message_type:
              type: string
              enum: [deployment.status.changed]
            data:
              type: object
              required: [deployment_status]
              properties:
                deployment_status:
                  $ref: "#/components/schemas/DeploymentStatus"
...
    OptimizationFeedbackEnvelope:
      allOf:
        - $ref: "#/components/schemas/CommonMessageEnvelope"
        - type: object
          required: [data]
          properties:
            message_type:
              type: string
              enum: [optimization.feedback.created]
            data:
              type: object
              required: [optimization_feedback]
              properties:
                optimization_feedback:
                  $ref: "#/components/schemas/OptimizationFeedback"
...
    FlowDelivery:
      type: object
      properties:
        correlation_id: {type: string}
        decision_id: {type: string}
        revision: {type: integer, enum: [1]}
        status: {type: string, enum: [PREPARED, SUBMISSION_UNKNOWN, SUBMITTED]}
        manifest_sha256: {type: string}
        destination_sha256: {type: string}
        request: {type: object, additionalProperties: true}
        deployment: {type: object, additionalProperties: true}
        metrics: {type: object, additionalProperties: true}
        limitations: {type: array, items: {type: string}}
        flow: {$ref: '#/components/schemas/AgentControlFlow'}
        updated_at: {type: string, format: date-time}
    AgentControlFlow:
      type: object
      required: [correlation_id, trace_id, state, updated_at]
      properties:
        correlation_id: {type: string}
        trace_id: {type: string}
        automation_run_id: {type: string}
        requested_decision_agent: {type: string}
        requested_operation_agent:
          type: string
          description: Operation Optimization Agent requested for the joined deployment status and Feedback.
        profile_id: {type: string}
        state:
          type: string
          enum:
            - WAITING_FOR_APPLICATION_CONTEXT
            - WAITING_FOR_RESOURCE_RECOMMENDATION
            - READY
            - AGENT_AUTHORIZATION_REJECTED
            - AGENT_EXECUTION_FAILED
            - AGENT_RESULT_REJECTED
            - DEPLOY_APPROVED
            - REJECTED
            - RETRY_REQUIRED
        application_context:
          $ref: "#/components/schemas/ApplicationContextEnvelope"
        resource_recommendation:
          $ref: "#/components/schemas/ResourceRecommendationEnvelope"
        agent_authorization:
          $ref: "#/components/schemas/AgentAuthorization"
        agent_execution:
          $ref: "#/components/schemas/DecisionAgentResult"
        decision:
          $ref: "#/components/schemas/AutomationDecision"
        deployment_plan:
          $ref: "#/components/schemas/AgentControlDeploymentPlan"
        guard:
          $ref: "#/components/schemas/AgentControlGuardResult"
        deployment_request:
          $ref: "#/components/schemas/DeploymentCreateRequestEnvelope"
        deployment_status:
          $ref: "#/components/schemas/DeploymentStatusEnvelope"
        optimization_feedback:
          $ref: "#/components/schemas/OptimizationFeedbackEnvelope"
        feedback_summary:
          $ref: "#/components/schemas/FeedbackSummary"
        operation_agent_execution:
          $ref: "#/components/schemas/OperationOptimizationResult"
        scaling_decision:
          $ref: "#/components/schemas/ScalingDecision"
        reasoning_comparison:
          $ref: "#/components/schemas/ReasoningComparison"
        updated_at: {type: string, format: date-time}
    AgentAuthorization:
      type: object
      required: [agent_name, capability, action, authorized, reason]
      properties:
        agent_name:
          type: string
          description: Selected configuration or Runtime Agent name
        capability:
          type: string
          enum: [ai_application_automation]
        action:
          type: string
          enum: [generate_deployment_decision]
        authorized: {type: boolean}
        reason: {type: string}
    DecisionAgentResult:
      type: object
      required: [agent_name, source, status, request_guard, result_guard, decision, latency_ms]
      properties:
        agent_name: {type: string}
        source: {type: string, enum: [configuration, runtime]}
        status: {type: string, enum: [completed, rejected, failed]}
        request_guard:
          $ref: "#/components/schemas/AgentControlGuardResult"
        result_guard:
          $ref: "#/components/schemas/AgentControlGuardResult"
        decision:
          $ref: "#/components/schemas/AutomationDecision"
        latency_ms: {type: integer, format: int64, minimum: 0}
        message: {type: string}
...
    AutomationDecision:
      type: object
      required:
        - decision_id
        - action
        - reason
        - reasoning_mode
        - confidence
        - created_at
      properties:
        decision_id: {type: string}
        action: {type: string, enum: [DEPLOY, REJECT, RETRY]}
        reason: {type: string}
        reasoning_mode: {type: string, enum: [rule_based]}
        confidence: {type: number, minimum: 0, maximum: 1}
        selected_candidate_id: {type: string}
        correction_request:
          $ref: "#/components/schemas/CorrectionRequest"
        created_at: {type: string, format: date-time}
    AgentControlDeploymentPlan:
      type: object
      required:
        - plan_id
        - profile_id
        - app_id
        - app_version
        - selected_candidate_id
        - target_runtime
        - desired_infrastructure
        - inference_configuration
        - reasoning_mode
        - created_at
      properties:
        plan_id: {type: string}
        profile_id: {type: string}
        app_id: {type: string}
        app_version: {type: string}
        selected_candidate_id: {type: string}
        target_runtime: {type: string, enum: [VM]}
        desired_infrastructure:
          $ref: "#/components/schemas/DesiredInfrastructure"
        inference_configuration:
          type: object
          additionalProperties: true
        resource_hints: {type: array, items: {type: string}}
        reasoning_mode: {type: string, enum: [rule_based]}
        created_at: {type: string, format: date-time}
    AgentControlGuardResult:
      type: object
      required: [status, checks]
      properties:
        status:
          type: string
          enum: [APPROVED, REJECTED, RETRY_REQUIRED]
        checks:
          type: array
          items:
            type: object
            required: [name, passed, reason]
            properties:
              name: {type: string}
              passed: {type: boolean}
              reason: {type: string}
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
    DeploymentCreateRequestEnvelope:
      allOf:
        - $ref: "#/components/schemas/CommonMessageEnvelope"
        - type: object
          required: [data]
          properties:
            message_type:
              type: string
              enum: [deployment.create.request]
            data:
              type: object
              required: [deployment_request]
              properties:
                deployment_request:
                  $ref: "#/components/schemas/AgentControlDeploymentRequest"
...
    FeedbackSummary:
      type: object
      required: [deployment_id, decision_id, deployment_state, success, cause, updated_at]
      properties:
        deployment_id: {type: string}
        decision_id: {type: string}
        deployment_state: {type: string}
        outcome: {type: string}
        success: {type: boolean}
        cause: {type: string}
        slo_violations: {type: array, items: {type: string}}
        metrics:
          $ref: "#/components/schemas/OptimizationMetrics"
        updated_at: {type: string, format: date-time}
    ScalingDecision:
      type: object
      required: [action, reason, current_replicas, desired_replicas, created_at]
      properties:
        action:
          type: string
          description: >
            Canonical scaling recommendation. KEEP preserves the observed
            replica count. Legacy NO_ACTION is accepted only as an Operation
            Agent proposal/result value and is normalized to KEEP before Guard
            validation; new responses use KEEP.
          enum: [KEEP, SCALE_OUT, SCALE_IN]
        reason:
          type: string
          minLength: 1
          maxLength: 8000
        current_replicas: {type: integer, minimum: 1}
        desired_replicas: {type: integer, minimum: 1}
        evidence:
          type: array
          items: {type: string}
        created_at: {type: string, format: date-time}
    OperationOptimizationResult:
      type: object
      required: [run_id, status]
      description: >
        Evidence returned by the Operation Optimization Agent after deployment
        status and optimization Feedback are joined. run_id is a deterministic,
        auditable operation execution identifier derived from the Flow and
        Feedback event. A completed result carries
        source, request_guard, result_guard, scaling_guard, and decision; all
        Guards remain outside the Agent. Terminal failed or rejected evidence
        may omit unavailable identity, Guard, latency, and decision fields and
        contains only terminal-safe evidence. Only a completed result with
        approved Guards can populate scaling_decision.
      properties:
        run_id: {type: string}
        agent_name:
          type: string
          description: Present when the selected Agent identity is known.
          example: OperationOptimizationAgent
        source:
          type: string
          enum: [configuration, runtime]
          description: Present for a resolved Agent.
        status: {type: string, enum: [completed, rejected, failed]}
        latency_ms:
          type: integer
          format: int64
          description: Present when execution latency is available.
        request_guard:
          $ref: "#/components/schemas/AgentControlGuardResult"
        result_guard:
          $ref: "#/components/schemas/AgentControlGuardResult"
        scaling_guard:
          $ref: "#/components/schemas/AgentControlGuardResult"
        decision:
          $ref: "#/components/schemas/ScalingDecision"
        message: {type: string, description: Terminal-safe execution summary.}
...
    ReasoningComparison:
      type: object
      required:
        - correlation_id
        - candidate_id
        - rule_based
        - simple_inference
        - validated_inference
        - agreement
        - created_at
      properties:
        correlation_id: {type: string}
        candidate_id: {type: string}
        rule_based:
          $ref: "#/components/schemas/ReasoningTrial"
        simple_inference:
          $ref: "#/components/schemas/ReasoningTrial"
        validated_inference:
          $ref: "#/components/schemas/ReasoningTrial"
        agreement:
          type: object
          required: [action_match, candidate_match]
          properties:
            action_match: {type: boolean}
            candidate_match: {type: boolean}
        created_at: {type: string, format: date-time}
...